Sebastian's branch

+++ mobile/TrainingPlan/Models/SurveyModels.swift
@@ -7,7 +7,7 @@ struct SurveyIn: Codable {
     let days_per_week: Int
     let days_of_week: [String]
     let most_time_day: String
-    let current_5k_fitness: Double
+    let current_5k_fitness: Int
     let major_injuries: String
     let most_recent_injury: String
 }
